refactored login routes to auth instead of api
