--- server/routes/index.js
@@ -1,11 +1,0 @@
-// const pluralize = require('pluralize');
-
-// TODO tailor to this project
-// app.use('/auth', require('./routes/user'));
-
-// let routearray = ['product', 'category', 'cart',
-//                     'lineitem', 'order', 'address',
-//                     'creditcard', 'review'];
-// routearray.forEach(item => {
-//     app.use(`/api/${pluralize(item)}`, require(`./routes/${item}`));
-// });
